+++ src/main/java/org/cicirello/sequences/SequenceReservoirSampler.java
@@ -61,6 +61,11 @@ public final class SequenceReservoirSampler implements SequenceSampler {
     this.r = new EnhancedRandomGenerator(seed);
   }
 
+  /** Constructs a sampler with a default source of randomness. */
+  public SequenceReservoirSampler() {
+    this.r = new EnhancedRandomGenerator();
+  }
+
   /**
    * package use only: creates an instance wrapping an existing EnhancedRandomGenerator
    *